#3 Refactor on message method to comply with message format update
The format for the message of the schedule/request was changed so that
the previous and new interval are included in the message instead of only
including the new interval. This allows the schedule handler to remove
unecessary intervals from its schedules map. This commit also change the
swich block to use the add to and remove from map functions added in the
previous commit.