#3 Refactor publishUpdatedSchedule() to publish to the correct topic
Although the getScheduleResponseTopic() method has been implemented a
while ago, the publishUpdatedSchedule() was still using the getStringDate()
from the filter file. This was resultin in the incorrect topic for dates which had
single digits days (ex: 2022-12-04). This commit change the implementation
of the publishUpdatedSchedule() function to use the
getScheduleResponseTopic() and removes the now unecessary export from
the filter export.